Rename multiple_comparrison to multiple_comparison
